Extract shared DAC codec builder and use torchaudio resample
